drm/xe: Remove pm_runtime lockdep
This lockdep was initially designed for the mem_access,
where the mem_access needed to run the resume sync from
the innerbounds and count the references.

With the runtime moving to the outer bounds of the driver
and the mem_access replaced by the pure rpm get/put
references, it is no longer needed and it is in a matter
of fact just splatting may false positives as the following:

[  200.615307]  Possible unsafe locking scenario:

[  200.621253]        CPU0                    CPU1
[  200.625804]        ----                    ----
[  200.630357]   lock(xe_pm_runtime_lockdep_map);
[  200.634821]                                lock(&power_domains->lock);
[  200.641378]                                lock(xe_pm_runtime_lockdep_map);
[  200.648372]   lock(&power_domains->lock);
[  200.652400]
                *** DEADLOCK ***

[  430.194948]  Possible unsafe locking scenario:

[  430.200895]        CPU0                    CPU1
[  430.205448]        ----                    ----
[  430.210000]   lock(&guc->submission_state.lock);
[  430.214638]                                lock(xe_pm_runtime_lockdep_map);
[  430.221630]                                lock(&guc->submission_state.lock);
[  430.228800]   lock((work_completion)(&sched->work_process_msg));
[  430.234834]
                *** DEADLOCK ***

In a matter of fact, there's actually a third lock that is not
in this picture:
spin_lock_irq(&dev->power.lock);
and INIT_WORK(&dev->power.work, pm_runtime_work);

The pm_callback_task will ensure that there's no recursive
calls of the resume function and it will increase the
reference counter anyway.

Then, the pm_runtime workqueue and spin locks will avoid that
any resume and suspend operations happens in parallel with
other resume and suspend operations.

With that, the only thing that we are actually doing here is
to wrongly train the lockdep, basically saying that we will
acquire some locks on resume and on suspend concurrently,
entirely ignoring its serialization and protection.

The above scenarios are simply not possible because there's
a serialization with the spin_lock_irq(&dev->power.lock)
before each operation. However we are telling the lockep
that the lock(xe_pm_runtime_lockdep_map) occurs before
the &dev->power.lock and lockdep is not capable to see
that other protection.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>