drm/i915/display: Introduce intel_display_guc_metrics
This is a generic component part of intel_display that collects
display information that could be used with GuC SLPC for a optimal
frequency selection.

This is also an experimental way to share the code with Xe without
a compact-headers, on an attempt to move towards the full detachment
of the intel_display from the i915.

Cc: Jani Nikula <jani.nikula@intel.com>
Cc: Lucas De Marchi <lucas.demarchi@intel.com>
Cc: Vinay Belgaumkar <vinay.belgaumkar@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>