drm/i915/display: Introduce Display Metrics info
Introduce a display metrics information through debugfs
for a better view of the vblank and page flips, in special
Async flips behavior.

There is currently an overall expectation that whenever
vblank_mode=0 is used with an graphics application, that
automatically async_flips are happening. However, while
implementing the Display Metrics for GuC SLPC, it was observed
that it is not necessarily true for many of the expected cases.

So, while the GuC SLPC side of the metrics doesn't get ready,
let's at least bring the debugfs view of it so we can work
to understand and fix any potential issue around our async vblanks.

Please notice that the every struct here follows exactly the
GuC shared data buffer, so the next step of the integration
would be smooth and almost transparent to this intel_metrics
on the display side.

v2: use Kzalloc instead of Kmalloc (Naveen)

Cc: Uma Shankar <uma.shankar@intel.com>
Cc: Vandita Kulkarni <vandita.kulkarni@intel.com>
Cc: Naveen Kumar <naveen1.kumar@intel.com>
Cc: Jani Nikula <jani.nikula@intel.com>
Cc: Vinay Belgaumkar <vinay.belgaumkar@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>