drm/xe: Relax runtime pm protection around VM
In the regular use case scenario, user space will create a
VM, and keep it alive for the entire duration of its workload.

For the regular desktop cases, it means that the VM
is alive even on idle scenarios where display goes off. This
is unacceptable since this would entirely block runtime PM
indefinitely, blocking deeper Package-C state. This would be
a waste drainage of power.

So, let's limit the protection only for the long running workloads,
which memory might be mapped and accessed during this entire
workload.

This indeed opens up a risk of use case without display, and
without long-running workload, where memory might be mapped
and accessed with direct read and write operations without
any gpu execution involved. Because of this, we are also
adding here, the extra protection for the special vm_op access
callback.

In the ideal case of the mmapped scenario of vm_ops, we would
also get references in the 'open' and 'mmap' callbacks, and
put it back on the 'close' callback, for a balanced case.
However, this would also block the regular desktop case, so
we are not doing this.

Cc: Thomas Hellström <thomas.hellstrom@linux.intel.com>
Cc: Lucas De Marchi <lucas.demarchi@intel.com>
Cc: Matthew Brost <matthew.brost@intel.com>
Cc: Francois Dugast <francois.dugast@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>