drm/xe/display: Kill crtc commit flush
This flush was needed in regular suspend cases in the past.
After the clean-up and reconciliation with the i915 it was
not needed anymore and removed. However this remained here
in the runtime suspend path.

However, the runtime pm flow ensures that there won't be any
flying or pending crtc commit when the runtime_suspend is
called. So this is not needed here. Clean it up.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>