drm/xe: Add quirk to block D3Cold if platform end in critical issues
D3Cold is a great power saving feature, that allows our GPUs to
drop to zero watts of consumption when in idle. It is already working
well, in most of the Intel discrete devices out there with Xe driver.

However, some systems are not behaving like expected at this time.

So, instead of disabling the D3Cold entirely for everyone while
the issue is investigated and debugged, let's flip the assumption,
leaving it enabled for everyone and adding to the list only the
ones where we are facing some blocking issues.

In these first cases, every transition from D3Cold to D0, the
upstream bridge fails to train the PCIe link even on slow speeds (Gen1)
and all the PCI devices underneath (audio and gpu) are entirely lost
with BAR reads only returning 0xffff.

This quirk only set the VRAM threshold limit to 0, forcing the driver
to disallow d3cold based on VRAM consumption, instead of entirely
removing the d3cold capability. This is intentional, so we have ways
of enabling it at runtime for experiments.

Link: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/4395
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>